Start cleaning this mess up
