+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: BuildPortals
-main: frahm.justin.mcplugins.buildportals.BuildPortals
+main: space.frahm.buildportals.BuildPortals
 version: 2.0.0
 softdepend:
   - Multiverse-Core
